Result of request now can be queried back to Unity; Extended count of unique classifications to 512; 16 bit classification is supported now.

+++ Octree.h
@@ -17,7 +17,7 @@ struct VertexData
 {
     glm::vec3 position;
     byte color[4];
-    byte classification;
+    int classification;
 };
 
 struct octreeSearchInfo
